auth_service/requirements.txt
flask
pyjwt
werkzeug
pyodbc
